+++ Engine/Source/Runtime/CoreUObject/Private/Serialization/AsyncLoadingThread.h
@@ -275,9 +275,9 @@ public:
 			int32 InPackagePriority,
 			const FLinkerInstancingContext* InstancingContext) override;
 
-	EAsyncPackageState::Type ProcessLoading(bool bUseTimeLimit, bool bUseFullTimeLimit, float TimeLimit) override;
+	EAsyncPackageState::Type ProcessLoading(bool bUseTimeLimit, bool bUseFullTimeLimit, double TimeLimit) override;
 
-	EAsyncPackageState::Type ProcessLoadingUntilComplete(TFunctionRef<bool()> CompletionPredicate, float TimeLimit) override;
+	EAsyncPackageState::Type ProcessLoadingUntilComplete(TFunctionRef<bool()> CompletionPredicate, double TimeLimit) override;
 
 	void FlushLoading(int32 PackageId) override;
